+++ lib/archive_today.rb
@@ -1,3 +1,6 @@
+require 'faraday'
+require 'faraday_middleware'
+require 'nokogiri'
 require 'archive_today/version'
 require 'archive_today/archiver'
 
@@ -5,8 +8,8 @@ module ArchiveToday
   class Error < StandardError; end
 
   class << self
-    def submit(url:, debug: false)
-      Archiver.new(url: url, debug: debug).submit
+    def capture(url:, debug: false)
+      Archiver.new(url: url, debug: debug).capture
     end
   end
 end