+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.2.0
+
+Also retrieve screenshot URL if available
+
+## Breaking Changes
+
+- Method renamed from `#submit` to `#capture`
+- Return type is now a hash with keys `:url` and `:screenshot_url`
+
 # 0.1.0
 
 Initial commit